requirements.txt
streamlit
langchain
langgraph
groq
dotenv
requests
grpcio
pinecone
feedparser
google-api-python-client
google-genai
numpy
scikit-learn
fitz